Python MicroService/requirements.txt
grpcio
grpcio-tools
scikit-learn
pdfminer